Projeto Finalizado

+++ src/main/java/com/dio/santander/bankline/api/model/Conta.java
@@ -2,14 +2,14 @@ package com.dio.santander.bankline.api.model;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import javax.persistence.Embedded;
+import java.math.BigDecimal;
 
 @Embeddable
 public class Conta {
     @Column(name = "conta_numero")
     private Long numero;
     @Column(name = "conta_saldo")
-    private Double saldo;
+    private BigDecimal saldo;
 
     public Conta() {
 
@@ -23,22 +23,22 @@ public class Conta {
         this.numero = numero;
     }
 
-    public Double getSaldo() {
+    public BigDecimal  getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(Double saldo) {
+    public void setSaldo(BigDecimal  saldo) {
         this.saldo = saldo;
     }
 
-    private Conta(Long numero, Double saldo) {
+    private Conta(Long numero, BigDecimal  saldo) {
         this.numero = numero;
         this.saldo = saldo;
     }
 
     public static class ContaBuild{
         private Long numero;
-        private Double saldo;
+        private BigDecimal  saldo;
 
         public ContaBuild() {
         }
@@ -46,7 +46,7 @@ public class Conta {
             this.numero=numero;
             return this;
         }
-        public ContaBuild saldo(Double saldo){
+        public ContaBuild saldo(BigDecimal  saldo){
             this.saldo=saldo;
             return this;
         }
